+++ src/main/java/kitchenpos/eatinorders/domain/Order.java
@@ -12,7 +12,7 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.persistence.Transient;
-import kitchenpos.eatinorders.todo.domain.OrderTable;
+import kitchenpos.eatinorders.todo.domain.ordertable.OrderTable;
 
 import java.time.LocalDateTime;
 import java.util.List;
