+++ src/main/java/kitchenpos/eatinorders/infra/JpaOrderTableRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.eatinorders.infra;
 
-import kitchenpos.eatinorders.todo.domain.OrderTable;
-import kitchenpos.eatinorders.todo.domain.OrderTableRepository;
+import kitchenpos.eatinorders.todo.domain.ordertable.OrderTable;
+import kitchenpos.eatinorders.todo.domain.ordertable.OrderTableRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
