+++ src/test/java/kitchenpos/eatinorders/application/OrderTableServiceTest.java
@@ -2,10 +2,10 @@ package kitchenpos.eatinorders.application;
 
 import kitchenpos.eatinorders.domain.OrderRepository;
 import kitchenpos.eatinorders.domain.OrderStatus;
-import kitchenpos.eatinorders.todo.domain.NumberOfGuests;
-import kitchenpos.eatinorders.todo.domain.OrderTable;
-import kitchenpos.eatinorders.todo.domain.OrderTableName;
-import kitchenpos.eatinorders.todo.domain.OrderTableRepository;
+import kitchenpos.eatinorders.todo.domain.ordertable.NumberOfGuests;
+import kitchenpos.eatinorders.todo.domain.ordertable.OrderTable;
+import kitchenpos.eatinorders.todo.domain.ordertable.OrderTableName;
+import kitchenpos.eatinorders.todo.domain.ordertable.OrderTableRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
